feat: CompanyService 인터페이스 분리 및 DEMO 파일 삭제 - CompanyService 인터페이스와 구현체 분리했습니다 - DEMO 파일 삭제했습니다
