--- demo/src/main/java/com/example/demo/domain/entity/User.java
@@ -1,10 +1,0 @@
-package com.example.demo.domain.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-
-@Entity
-public class User {
-    @Id
-    private Long id;
-}
