--- demo/src/main/java/com/example/demo/global/configuration/RedisConfig.java
@@ -1,5 +1,0 @@
-package com.example.demo.global.configuration;
-
-public class RedisConfig {
-
-}
