--- demo/src/main/java/com/example/demo/global/exception/ExceptionCode.java
@@ -1,12 +1,0 @@
-package com.example.demo.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-
-    String name();
-
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-}
