--- demo/src/main/java/com/example/demo/infrastructure/jpa/querydsl/UserQueryDslRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.infrastructure.jpa.querydsl;
-
-import com.example.demo.domain.entity.User;
-import java.util.Optional;
-
-
-public interface UserQueryDslRepository {
-
-    Optional<User> findByName(String name);
-}
