--- demo/src/main/java/com/example/demo/infrastructure/jpa/querydsl/UserQueryDslRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.example.demo.infrastructure.jpa.querydsl;
-
-import com.example.demo.domain.entity.User;
-import java.util.Optional;
-
-public class UserQueryDslRepositoryImpl implements UserQueryDslRepository {
-
-    @Override
-    public Optional<User> findByName(String name) {
-        return Optional.empty();
-    }
-}
